#21 - config hook was unecessary
