--- src/hooks/useConfig.tsx
@@ -1,19 +1,0 @@
-import { ServerConfig, ServerConfigContainer } from 'containers/ConfigContainer'
-import { useEffect, useState } from 'react'
-
-const useConfig = () => {
-  const { getConfig } = ServerConfigContainer.useContainer()
-  const [appConfig, setAppConfig] = useState<ServerConfig>()
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const data = await getConfig()
-      setAppConfig(data)
-    }
-    fetchData()
-  }, [getConfig])
-
-  return appConfig
-}
-
-export default useConfig
