+++ src/types/custom_types.ts
@@ -255,9 +255,3 @@ export interface CommandParams {
 export interface TableInterface {
   parentState: TableState
 }
-
-export interface VersionConfig {
-  beer_garden_version: string
-  current_api_version: string
-  supported_api_versions: [string]
-}
